requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
blas=1.0=mkl
brotli-python=1.0.9=py39hcec6c5f_8
bzip2=1.0.8=h6c40b1e_6
ca-certificates=2024.11.26=hecd8cb5_0
certifi=2024.8.30=py39hecd8cb5_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
cloudpickle=3.0.0=py39hecd8cb5_0
farama-notifications=0.0.4=py39hecd8cb5_0
filelock=3.13.1=py39hecd8cb5_0
freetype=2.12.1=h60636b9_2
gmp=6.2.1=he9d5cce_3
gmpy2=2.1.2=py39hd5de756_0
gym=0.26.1=py39h2a79d86_0
gym-notices=0.0.8=pyhd8ed1ab_0
gymnasium=0.28.1=py39hecd8cb5_1
idna=3.7=py39hecd8cb5_0
importlib-metadata=8.5.0=py39hecd8cb5_0
importlib_metadata=8.5.0=hd3eb1b0_0
intel-openmp=2023.1.0=ha357a0b_43548
jax-jumpy=1.0.0=py39hecd8cb5_0
jinja2=3.1.4=py39hecd8cb5_1
lcms2=2.16=ha2f27b4_0
lerc=4.0.0=hb486fe8_0
libcxx=19.1.4=hf95d169_0
libdeflate=1.22=h00291cd_0
libffi=3.4.4=hecd8cb5_1
libjpeg-turbo=3.0.3=h46256e1_0
libpng=1.6.44=h4b8f8c9_0
libsqlite=3.47.0=h2f8c449_1
libtiff=4.7.0=h583c2ba_1
libwebp-base=1.4.0=h10d778d_0
libxcb=1.17.0=hf1f96e2_0
libzlib=1.3.1=hd23fc13_2
markupsafe=2.1.3=py39h6c40b1e_0
mkl=2023.1.0=h8e150cf_43560
mkl-service=2.4.0=py39h6c40b1e_1
mkl_fft=1.3.8=py39h6c40b1e_0
mkl_random=1.2.4=py39ha357a0b_0
mpc=1.1.0=h6ef4df4_1
mpfr=4.0.2=h9066e36_1
mpmath=1.3.0=py39hecd8cb5_0
ncurses=6.4=hcec6c5f_0
networkx=3.2.1=py39hecd8cb5_0
numpy=1.26.4=py39h827a554_0
numpy-base=1.26.4=py39ha186be2_0
openjpeg=2.5.2=hbf2204d_0
openssl=3.4.0=hd471939_0
pcre=8.45=h23ab428_0
pillow=11.0.0=py39h6cf2171_0
pip=24.2=py39hecd8cb5_0
pthread-stubs=0.3=hdd91f34_1
pygame=2.1.3=pypi_0
pysocks=1.7.1=py39hecd8cb5_0
python=3.9.18=h7a9c478_1_cpython
python_abi=3.9=5_cp39
pytorch=2.2.2=py3.9_0
pyyaml=6.0.2=py39h46256e1_0
readline=8.2=hca72f7f_0
requests=2.32.3=py39hecd8cb5_1
setuptools=75.1.0=py39hecd8cb5_0
swig=4.0.2=he9d5cce_4
sympy=1.13.2=py39hecd8cb5_0
tbb=2021.8.0=ha357a0b_0
tk=8.6.13=h1abcd95_1
torchaudio=2.2.2=py39_cpu
torchvision=0.17.2=py39_cpu
typing-extensions=4.11.0=py39hecd8cb5_0
typing_extensions=4.11.0=py39hecd8cb5_0
tzdata=2024b=h04d1e81_0
urllib3=2.2.3=py39hecd8cb5_0
wheel=0.44.0=py39hecd8cb5_0
xorg-libxau=1.0.11=h00291cd_1
xorg-libxdmcp=1.1.5=h00291cd_0
xz=5.4.6=h6c40b1e_1
yaml=0.2.5=haf1e3a3_0
zipp=3.21.0=py39hecd8cb5_0
zstd=1.5.6=h915ae27_0
